+++ pg_bm25/Cargo.toml
@@ -18,7 +18,6 @@ pg15 = ["pgrx/pg15", "pgrx-tests/pg15"]
 pg16 = ["pgrx/pg16", "pgrx-tests/pg16"]
 pg_test = []
 telemetry = ["shared/telemetry"]
-icu = ["rust_icu_ubrk"]
 
 [dependencies]
 anyhow = { version = "1.0.79", features = ["backtrace"] }
@@ -32,18 +31,9 @@ indexmap = "2.1.0"
 interprocess = "1.2.1"
 json5 = "0.4.1"
 libc = "0.2.152"
-lindera-core = "0.27.1"
-lindera-dictionary = "0.27.1"
-lindera-tokenizer = { version = "0.27.1", features = [
-  "cc-cedict-compress",
-  "cc-cedict",
-  "ipadic-compress",
-  "ipadic",
-  "ko-dic-compress",
-  "ko-dic",
-] }
 memoffset = "0.9.0"
 once_cell = "1.18.0"
+tokenizers = { version = "0.1.0", path = "../tokenizers" }
 pgrx = "=0.11.2"
 reqwest = "0.11.22"
 rustc-hash = "1.1.0"
@@ -69,7 +59,3 @@ rstest = "0.18.2"
 shared = { path = "../shared", features = ["fixtures"] }
 sqlx = { version = "0.7.3", features = ["postgres", "runtime-async-std", "time", "bigdecimal", "uuid"] }
 tempfile = "3.9.0"
-
-[dependencies.rust_icu_ubrk]
-version = "4.2.3"
-optional = true
