+++ pg_bm25/src/lib.rs
@@ -5,7 +5,6 @@ mod index;
 mod postgres;
 mod query;
 mod schema;
-mod tokenizers;
 mod writer;
 
 #[cfg(test)]
