+++ pg_bm25/src/schema/config.rs
@@ -96,3 +96,101 @@ where
         Ok(output)
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use rstest::*;
+    use tantivy::schema::{JsonObjectOptions, NumericOptions, TextOptions};
+
+    use crate::schema::SearchFieldConfig;
+
+    #[rstest]
+    fn test_search_text_options() {
+        let json = r#"{
+            "indexed": true,
+            "fast": false,
+            "stored": true,
+            "fieldnorms": true,
+            "type": "default",
+            "record": "basic",
+            "normalizer": "raw"
+        }"#;
+        let config: serde_json::Value = serde_json::from_str(json).unwrap();
+        let search_text_option: SearchFieldConfig =
+            serde_json::from_value(serde_json::json!({"Text": config})).unwrap();
+        let expected: TextOptions = search_text_option.into();
+
+        let text_options: TextOptions = SearchFieldConfig::default_text().into();
+        assert_eq!(expected.is_stored(), text_options.is_stored());
+        assert_eq!(
+            expected.get_fast_field_tokenizer_name(),
+            text_options.get_fast_field_tokenizer_name()
+        );
+
+        let text_options = text_options.set_fast(Some("index"));
+        assert_ne!(expected.is_fast(), text_options.is_fast());
+    }
+
+    #[rstest]
+    fn test_search_numeric_options() {
+        let json = r#"{
+            "indexed": true,
+            "stored": true,
+            "fieldnorms": false,
+            "fast": true
+        }"#;
+        let config: serde_json::Value = serde_json::from_str(json).unwrap();
+        let expected: SearchFieldConfig =
+            serde_json::from_value(serde_json::json!({"Numeric": config})).unwrap();
+        let int_options: NumericOptions = SearchFieldConfig::default_numeric().into();
+
+        assert_eq!(int_options, expected.into());
+    }
+
+    #[rstest]
+    fn test_search_boolean_options() {
+        let json = r#"{
+            "indexed": true,
+            "stored": true,
+            "fieldnorms": false,
+            "fast": true
+        }"#;
+        let config: serde_json::Value = serde_json::from_str(json).unwrap();
+        let expected: SearchFieldConfig =
+            serde_json::from_value(serde_json::json!({"Boolean": config})).unwrap();
+        let int_options: NumericOptions = SearchFieldConfig::default_numeric().into();
+
+        assert_eq!(int_options, expected.into());
+    }
+
+    #[rstest]
+    fn test_search_jsonobject_options() {
+        let json = r#"{
+            "indexed": true,
+            "fast": false,
+            "stored": true,
+            "expand_dots": true,
+            "type": "default",
+            "record": "basic",
+            "normalizer": "raw"
+        }"#;
+        let config: serde_json::Value = serde_json::from_str(json).unwrap();
+        let search_json_option: SearchFieldConfig =
+            serde_json::from_value(serde_json::json!({"Json": config})).unwrap();
+        let expected: JsonObjectOptions = search_json_option.into();
+
+        let json_object_options: JsonObjectOptions = SearchFieldConfig::default_json().into();
+        assert_eq!(expected.is_stored(), json_object_options.is_stored());
+        assert_eq!(
+            expected.get_fast_field_tokenizer_name(),
+            json_object_options.get_fast_field_tokenizer_name()
+        );
+        assert_eq!(
+            expected.is_expand_dots_enabled(),
+            json_object_options.is_expand_dots_enabled()
+        );
+
+        let text_options = json_object_options.set_fast(Some("index"));
+        assert_ne!(expected.is_fast(), text_options.is_fast());
+    }
+}
