+++ tokenizers/Cargo.toml
@@ -1,0 +1,33 @@
+[package]
+name = "tokenizers"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[features]
+icu = ["rust_icu_ubrk"]
+
+[dependencies]
+lindera-core = "0.27.1"
+lindera-dictionary = "0.27.1"
+lindera-tokenizer = { version = "0.27.1", features = [
+  "cc-cedict-compress",
+  "cc-cedict",
+  "ipadic-compress",
+  "ipadic",
+  "ko-dic-compress",
+  "ko-dic",
+] }
+once_cell = "1.18.0"
+serde = "1.0.188"
+serde_json = "1.0.105"
+tantivy = { git = "https://github.com/paradedb/tantivy.git", package = "tantivy", rev = "e678820" }
+tracing = "0.1.40"
+
+[dependencies.rust_icu_ubrk]
+version = "4.2.3"
+optional = true
+
+[dev-dependencies]
+rstest = "0.18.2"
